Add Support for IMAP node.
The 'IMAP' node is used to map voxel indexes (the 'i' value on each
Voxel structure) to the color palette in a way that is *not* simply
value-to-value.

This structure has the following behavior if it appears in a file:
https://github.com/ephtracy/voxel-model/issues/19#issuecomment-739324018

When importing a file which contains this, imported models will not
match the expected values in the palette unless the IMAP is respected.